// viasoft.licensing.customerlicensing-frontend/src/client/customer-licensing/model/organizationUnitEnvironment.ts
/**
 * Viasoft.Licensing.LicensingManagement
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrganizationUnitEnvironment { 
    name: string;
    description?: string | null;
    isProduction?: boolean;
    isMobile?: boolean;
    isWeb?: boolean;
    isDesktop?: boolean;
    databaseName?: string | null;
    organizationUnitId?: string;
    isActive?: boolean;
    id?: string;
    desktopDatabaseVersion: string;
}